+++ winter-compiler/src/main/java/io/jentz/winter/compiler/InjectorModel.kt
@@ -4,15 +4,20 @@ import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.asClassName
 import javax.lang.model.element.TypeElement
 
-class InjectorModel(typeElement: TypeElement) {
+class InjectorModel(typeElement: TypeElement, superClassWithInjector: TypeElement?) {
 
     val typeName = typeElement.asClassName()
 
-    val generatedClassName = ClassName(
-            typeName.packageName,
-            "WinterMembersInjector_${typeName.simpleNames.joinToString("_")}"
-    )
+    val generatedClassName = generatedClassNameForClassName(typeName)
+
+    val superclassInjectorClassName = superClassWithInjector
+        ?.let { generatedClassNameForClassName(it.asClassName()) }
 
     val targets: MutableSet<InjectTargetModel> = mutableSetOf()
+
+    private fun generatedClassNameForClassName(name: ClassName) = ClassName(
+        name.packageName,
+        "${name.simpleNames.joinToString("_")}_WinterMembersInjector"
+    )
 
 }